intel_th: Use saved nwsa only when msc is disabled
Use the saved nwsa only when msc is disabled, otherwise use the value
provided by MSCxNWSA register.
The nswa (next window start address) backup was introduced to have a
valid starting window for post capture dumping in case of in memory
tracing, but a regresion was introduced for "online" memory tracing
(eg. dvc).

Signed-off-by: Traian Schiau <traianx.schiau@intel.com>