REVERTME: ASoC: Intel: CNL: Load firmware in dsp_init.
This patch loads the DSP firmware as a part of dsp_init
instead of dsp_fw init.  SoundWire master controller
doesnt get register if the DSP firmware is loaded
as part of fw_init, since for master controller registration
DSP power up is must, which is part of fw_load function.

Change-Id: I37edac4043abca18c189b035cb1e60194e33d79d
Signed-off-by: Guneshwor Singh <guneshwor.o.singh@intel.com>