[REVERTME]ASoC: Intel: CNL: Define DPN package for SDW1.
As DPN package entries are not defined in the BIOS for SoundWire
device 1(SDW1), the capability data-port-type can not be mapped
for SDW1.

As a workaround, hard code data-port-type for SDW1.
Workaround is not required for other devices, since
BIOS has entry for them.

This patch can be reverted once BIOS has entry for SDW1.

Change-Id: I15b0ec44731feb1f1f45466a0f9a9426d31a8dee
Signed-off-by: Anil Bhawangirkar <anil.k.bhawangirkar@intel.com>