README.internal: Add documentation for sim, emulation and internal patches
We currently use the SIM tag for simulation patches but this is not described
in README.internal. This patch documents this and adopts the convention
drm/i915/sim instead of the tag.

Emulation support will need to be added to the internal tree. For these patches
a drm/i915/emu label is introduced.

A statement on the intent to upstream none sim and emu patches once no longer
embargoed is added.

The !UPSTREAM tag is changed to NOT_UPSTREAM, as the ! is dropped in the file
names of patches generated from git.

Signed-off-by: Kelvin Gardiner <kelvin.gardiner@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>