!UPSTREAM drm/i915/icl: assume PCH_ICP and that DDI ports are present
We don't have proper support for the ICL PCH or DDI port detection, and
as a consequence we're not even initializing any connectors. Assume
that we have PCH_CNP and that ports A-E are present so that Simics at
least tries to initialize something.

We'll replace this patch with appropriate code when the developers
submit their work.

v2:
- We can also init port F now.
- It's already good enough to assume PCH_ICP.
- We need to split this patch into two actual patches.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>