drm/i915/icl: implement the display init/uninit sequences
This code is similar enough to the CNL code that I considered just
adding ICL support to the CNL function, but I think it's still
different enough, and having a function specific to ICL allows us to
more easily adapt code in case the spec changes more later.

We're still missing the power wells and the mbus code, so leave thoese
pieces with WARN_ON(1) while they're not here yet.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>