drm/i915/icl: add basic support for the ICL clocks
This commit introduces the definitions for the ICL clocks and adds the
basic functions to the shared DPLL framework. It adds code for the
Enable and Disable sequences for the Combo, TBT and MG PLLs, but it
does not have the code to compute the actual PLL values, which are
marked as TODO comments and should be introduced as separate commits.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>