x86/perf/intel: Support resetting metrics explicitely
The PERF_EVENT_IOC_RESET ioctl currently only resets the software count.
For TopDown it is useful to also reset the underlying hardware
counter because it controls additional hidden counters. Add
the necessary callbacks that a PMU driver can do that.

This allows RDPMC users of the Metrics counter to reset their
base line.

Add perf infrastructure and the necessary callbacks to support
that.

Signed-off-by: Andi Kleen <ak@linux.intel.com>