mei: dal: add test module
DAL test module allows the user space to test the DAL kernel space api.
It exposes a character device to the user space, and calls DAL api's
according to the protocol which is defined in kdi_cmd_defs.h header file.
This patch adds the code of DAL test module.

Change-Id: Id7ec2e7d3f22b07c41941dc5bada6edaf16e893b
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>
Signed-off-by: Yael Samet <yael.samet@intel.com>