drm/i915/icl: Enum for Type-C ports
Type-C ports are named with numeric suffixes - Type-C 1, Type-C 2, ...
or TC1, TC2, ... and have separate interrupt registers. As the number of
non-Type-C (combo) ports is platform dependent, we need a way to convert
enum port values for Type-C DDI ports to appropriate offsets into Type-C
specific registers. So create a new enum tc_port for Type-C ports and a
function to map enum port to enum tc_port depending on the platform.

Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Cc: Jani Nikula <jani.nikula@intel.com>
Reviewed-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>