drm/i915/icl/guc: Introduce new GuC interrupt register
Gen11 adds new Host2GuC interrupt register. It changed from previous
GENs and it is more flexible now. The host can write to this register
any 32-bit payload, which results in an interrupt to the GuC, which
can then retrieve this additional payload directly from the register
(instead of having a single bit for triggering the interrupt like before).
As current GuC firmware (16.75) ignores the payload for now, we just
write a 1 to bit 0 like before.

v6: rebased and split (Michal)
v7: rebased again (Michal)

Bspec: 21043
Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
Cc: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Oscar Mateo <oscar.mateo@intel.com>
Cc: Michel Thierry <michel.thierry@intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>