net: stmmac: solve error message on debugFS when waking up from sleep
DebugFS is not removed when system goes to sleep but is initialized
when waking up from sleep. This causes debugFS to fail to initialize
when system wakes up, which prompts an error message.

This problem can be solved by moving debugFS initialization out from
stmmac_hw_setup(), which is called by stmmac_resume(). By doing so,
debugFS initialization will not occur when system wakes up.

Signed-off-by: Ooi, Joyce <joyce.ooi@intel.com>
Reviewed-by: Zhang, Baoli <baoli.zhang@intel.com>