net: ethtool: introduce hardware tunable parameter for IEEE802.1 Qbu
'tx-fpe-afsz' is meant to change the non-final fragment size of preempted
frame. A smaller value here means express frame will be transmitted by
express MAC as soon as preempted MAC (pMAC) has collected the configured
fragment size, default to 64-byte fragment.

'tx-fpe-hadv' is meant to specify the time (in nano-seconds) for pMAC to
switch from release to hold state in advance.

'tx-fpe-radv' is meant to specify the time (in nano-seconds) for pMAC to
switch from hold to release state in advance.

Signed-off-by: Ong Boon Leong <boon.leong.ong@intel.com>