drm/i915/icl/guc: Capture & release GuC interrupts properly for Icelake
With the new interrupt repartitioning in Gen11, GuC controls by itself the interrupts
it receives, so steering bits and registers have been defeatured. Being this the case,
when the GuC is in control of submissions we won't know what to do with the ctx switch
interrupt in the KMD, so disable it.

Bspec: 12609
Bspec: 10800
Bspec: 10932
Bspec: 10934
Bspec: 9517

v2: For the GuC, we still need ARAT expired interrupt unmasked (Sagar)

v3: Remove FIXME from the commit message (Daniele)

Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: Michal Wajdeczko <michal.wajdeczko@intel.com>
Cc: Sagar Arun Kamble <sagar.a.kamble@intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: Oscar Mateo <oscar.mateo@intel.com>