drm/i915/icl/guc: Add support for Icelake GuC IDs based on engine class and instance
Starting from Gen11, the ID to be provided to GuC needs to
contain the class in bits [0,2] and the instance in bits [3,6].

NOTE: this patch breaks pointer dereferences in some existing guc
functions that use the guc_id to dereference arrays. These functions are
not applicable to Gen11 and new functions will be pushed in follow up
patches, but the ID support is a requirement for those. Since the GuC FW has
not been defined yet for any GEN11 platform any GuC-related operation will be
aborted so we won't use the broken code and won't have any real issue.

Bspec: 20944

v2: Do not cause conflicts with the existing HACK, so that we can drop it
    more easily (Daniele)

v3: Rebased

v4:
  - Correct Bspec number (Michal)
  - Define GEN11_GUC_ENGINE_ALL_INSTANCES (Michal)
  - Clean GEN11_GUC_ENGINE_* a little bit

v5: Clarify the use of bit 7 (Daniele)

v6: Rebased (Michel)

v7: fix GEN11_GUC_ENGINE_CLASS_MASK, rename GEN11_GUC_ENGINE_* to
    GEN11_GUC_ID_TO_ENGINE_* (Daniele)

Cc: Michal Wajdeczko <michal.wajdeczko@intel.com>
Cc: Tomasz Lis <tomasz.lis@intel.com>
Signed-off-by: Ceraolo Spurio, Daniele <daniele.ceraolospurio@intel.com>
Signed-off-by: Oscar Mateo <oscar.mateo@intel.com>
Signed-off-by: Michel Thierry <michel.thierry@intel.com>