drm/i915/icl: implement icl_digital_port_connected()
Do like the other functions and check for the ISR bits. We have plans
to add a few more checks in this code in the next patchs, that's why
it's a little more verbose than it could be.

v2: [Rodrigo: pass encoder and check for hdp_pin instead of ports,
    following CNL patch]

v3: Fix macro order and alignment (Lucas)

Cc: Animesh Manna <animesh.manna@intel.com>
Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>