PCI: hotplug: Drop checking of PCI_BRIDGE_CONTROL in *_unconfigure_device()
During PCIe surprise hot-unplug the device is not accessible anymore and
register reads return 0xffffffff. When that happens pciehp_unconfigure_device()
may inadvertently think the device below the bridge may be a display
device of somesort as reading PCI_BRIDGE_CONTROL register also returns
0xff. This results failure of the remove operation:

  pciehp 0000:00:1c.0:pcie004: Slot(0): Link Down
  pciehp 0000:00:1c.0:pcie004: Slot(0): Card present
  pciehp 0000:00:1c.0:pcie004: Cannot remove display device 0000:01:00.0

Because of this the hierarchy is left untouched preventing further
hotplug operations.

Now, it is not clear why the check is there in the first place and why
we would like to prevent removing a bridge if it has PCI_BRIDGE_CTL_VGA
set. In case of PCIe surprise hot-unplug, it would not even be possible
to prevent the removal.

Given this and the issue described above, I think it makes sense to drop
the whole PCI_BRIDGE_CONTROL check from pciehp_unconfigure_device().
While there do the same for shpchp_configure_device() based on the same
reasoning and the fact that the same bug might trigger in standard PCI
hotplug as well.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>