drm/i915/tgl/guc: Distributed Doorbells
GuC doorbells are now split across different SQIDI (based on GT model),
and therefore the doorbell cache selection cannot be decided by us.

The doorbell cacheline will now be chosen between HW/GuC and returned to
us after a successful "acquire doorbell" command, embedded in the returned
action status dword. Then we only need to multiply this cacheline by the
cacheline size.

Doorbells need to be allocated widespread across different SQIDIs as this
claims to be a HW performance benefit, and it has been decided to use a
simple round robin technique (the algorithm must be the same in the
driver and the firmware code). The SQIDI selection will always start
from SQIDI0 and move on until the max SQIDI supported, once there, it
will wrap to SQIDI0.

Additionally we may end up having to look at other SQIDIs if one of them
runs out of doorbells (although we try to keep the doorbell assignment
balanced, after some deallocations, it could be the case that one of the
SQIDIs is full while the other has space left).

Note that although we no longer control the selection of cacheline, none
of the Gen12 defined skus ever set the highest bit. This means that the
doorbell cacheline will always be in the lower 2KB, and we can place the
application process descriptor in the upper half of the page.

v2: rebased, using updated guc->send() API (Michal)
v3: rebased.
v4: There is no need to update db status nor cookie after db allocation,
the fw should be doing this (assert the db status instead); rename the
function that updates the doorbell in proxy stage; symmetrical assert check
in legacy_create_doorbell. (Daniele)
v5: rebased.
v6: rebased.
v7: rebased.
v8: rebased.

Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
Signed-off-by: Michel Thierry <michel.thierry@intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>