NOT_UPSTREAM: drm/i915/guc/tgl: GuC semaphore feature flags
To handle staged additions to the models from HW/fulsim teams the GuC FW
gates all new HW features behind feature flags.

This patch add the feature flags for the new token-based semaphore
mechanism. We need to enable the new mechanism as the CSB changes are
part of it. Those changes landed in HW/fulsim many months ago so there
should be no risk of anyone running on models that don't have the support.

The flags will be removed from GuC FW after RTL freeze and this patch
can be dropped at that point.

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: Michel Thierry <michel.thierry@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>