NOT_UPSTREAM: drm/i915/tgl/guc: enforce TGL FW == ICL FW
The ICL and TGL GuC FWs currently have the same interface and are
following the same version scheme. If we take a new major FW version
(which comes with an interface update) for one of the 2 platforms we
need to make sure that the other one is in sync as well, so add a check
for that.

This patch can be dropped once the ICL FW is frozen and 2 FWs start
to diverge.

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: Michel Thierry <michel.thierry@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>