ethtool: added launch time fetch & expiry offset in hw tunable
'tx-tbs-leos' is the max amount of time of a frame is eligible for a
launch, starting from the time the frame becomes eligible for launch
(in nano-seconds).

'tx-tbs-legos' is the GCL Slot Number (GSN) offset (multiple of
cycle-time). This offset is only used when the launch time is set
as EST/Offset mode. A frame is eligible to launch before the
GSN offset + launch expiry offset.

'tx-tbs-estm' is meant for setting time based scheduling mode
to absolute / est mode. '0' means absolute mode and other value
mean est mode.

'tx-tbs-ftos' is meant for setting the dma fetch time offset
before the packet launch (in nano-seconds).

'tx-tbs-fgos' is meant for setting the dma fetch GSN offset
before the packet launch. Unit is integer.

Signed-off-by: Weifeng Voon <weifeng.voon@intel.com>
Signed-off-by: Kweh Hock Leong <hock.leong.kweh@intel.com>
Signed-off-by: Ong Boon Leong <boon.leong.ong@intel.com>
Reviewed-by: Baoli Zhang <baoli.zhang@intel.com>