drm/i915/icl: Mind the SFC units when resetting VD or VEBox engines
SFC units are shared between VD and VEBoxes. They also happen to have
separate reset bits. So, whenever we want to reset one or more of the
media engines, we have to make sure the SFCs do not change owner in the
process and, if this owner happens to be one of the engines being reset,
we need to reset the SFC as well.

This happens in 4 steps:

1) Tell the engine that a software reset is going to happen. The engine
will then try to force lock the SFC (if currently locked, it will
remain so; if currently unlocked, it will ignore this and all new lock
requests).

2) Poll the ack bit to make sure the hardware has received the forced
lock from the driver. Once this bit is set, it indicates SFC status
(lock or unlock) will not change anymore (until we tell the engine it
is safe to unlock again).

3) Check the usage bit to see if the SFC has ended up being locked to
the engine we want to reset. If this is the case, we have to reset
the SFC as well.

4) Unlock all the SFCs once the reset sequence is completed.

Obviously, if we are resetting the whole GPU, we don't have to worry
about all of this.

BSpec: 10989
BSpec: 10990
BSpec: 10954
BSpec: 10955
BSpec: 10956
BSpec: 19212

v2:
  - Added BSpec reference to GDRST (Michel)
  - Use () in macros that take instances (Michel)
  - No need for sfc_reset_bit in unlock (Michel)
  - New line between GEN11_GRDOM_SFC1 and SFC_RESET_BIT (Michel)

Signed-off-by: Oscar Mateo <oscar.mateo@intel.com>
Cc: Michel Thierry <michel.thierry@intel.com>
Signed-off-by: Michel Thierry <michel.thierry@intel.com>