drm/i915/guc: Update the reset-engine H2G command to support Gen11
From Gen11 onwards, not only the order and meaning of most the
parameters is different compared to the pre-Gen11 interface, but also
the firmware will send a G2H ENGINE_RESET_COMPLETE message (with the
engine's guc_class in data[2]) to confirm that the reset has been
completed.

As first steps, update the H2G command and add helper functions to flag
when an engine is under reset and wait_for the flag to be cleared.

The reset_complete handling is added in a subsequent patch.

Requires guc fw v25.161+.

Signed-off-by: Michel Thierry <michel.thierry@intel.com>
Cc: Oscar Mateo <oscar.mateo@intel.com>
Cc: Vinay Belgaumkar <vinay.belgaumkar@intel.com>
Tested-by: Vinay Belgaumkar <vinay.belgaumkar@intel.com>
Acked-by: Oscar Mateo <oscar.mateo@intel.com>