[REVERTME] ASoC: Intel: CNL: Fix for the firmware redownload failure on ICL
There is a bug in hda uncaching mechanism due to which seecond time
firmware download was failing in the ICL platform. In order to get
around this, flush_cache_range() is used as WA to maintain the
integrity of firmware. This patch will be reverted, once the proper
fix is available

Change-Id: I2efdfe82dc302e8a1b623c905da2ea08084d8074
Signed-off-by: R, Dharageswari <dharageswari.r@intel.com>
Signed-off-by: Pradeep Tewani <pradeep.d.tewani@intel.com>
Reviewed-on: 
Reviewed-by: Kale, Sanyog R <sanyog.r.kale@intel.com>
Reviewed-by: Prakash, Divya1 <divya1.prakash@intel.com>
Reviewed-by: Prusty, Subhransu S <subhransu.s.prusty@intel.com>
Tested-by: Madiwalar, MadiwalappaX <madiwalappax.madiwalar@intel.com>